requirements.txt
python-dotenv==1.0.1

langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.41
langchain-text-splitters==0.2.4


langchain-openai==0.1.22
openai==1.40.6

chromadb==0.5.5

beautifulsoup4==4.12.3
lxml==5.2.2
pypdf==4.3.1

streamlit==1.36.0
